Create load_cell_hx711_test.ino
here is a simple test to see if our weight sensor (load cell and hx711) is able to read a weight. If weight is detected it should print to serial monitor "weight was detected." "no weight detected." if not. We should also see changes in weight printed as items are placed or removed from the scale.